Script: modify how Runestone Python trace routines are organized (PR #1746)
