+++ pretext/README.md
@@ -39,8 +39,8 @@ making, placing, and employing a customized version.
 
 **NEVER EDIT THE ORIGINAL VERSION OF** `pretext.cfg`
 
-`pg_encoder.py`, `pg_logger.py`
--------------------------------
+`runestone`
+-----------
 
-Modules from Runestone to support the formulation of trace data for
-Python programs to poer CodeLens interactive programs written in Python.
+Selected modules from RunestoneComponents to support the formulation of trace data for
+Python programs to CodeLens interactive programs written in Python.
